+++ jax/_src/numpy/lax_numpy.py
@@ -1966,6 +1966,27 @@ empty_like = zeros_like
 empty = zeros
 
 
+# General np.from* style functions mostly delegate to numpy.
+
+@_wraps(np.frombuffer)
+def frombuffer(buffer, dtype=float, count=-1, offset=0):
+  return asarray(np.frombuffer(buffer=buffer, dtype=dtype, count=count, offset=offset))
+
+
+@_wraps(np.fromfunction)
+def fromfunction(function, shape, *, dtype=float, **kwargs):
+  shape = core.canonicalize_shape(shape, context="shape argument of jnp.fromfunction()")
+  for i in range(len(shape)):
+    in_axes = [0 if i == j else None for j in range(len(shape))]
+    function = jax.vmap(function, in_axes=tuple(in_axes[::-1]))
+  return function(*(arange(s, dtype=dtype) for s in shape), **kwargs)
+
+
+@_wraps(np.fromstring)
+def fromstring(string, dtype=float, count=-1, *, sep):
+  return asarray(np.fromstring(string=string, dtype=dtype, count=count, sep=sep))
+
+
 @_wraps(np.eye)
 def eye(N, M=None, k=0, dtype=None):
   lax_internal._check_user_dtype_supported(dtype, "eye")